# api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "ExchangeIn": {
        "additionalProperties": false,
        "description": "Canje del codigo de un solo uso por la sesion (US-405). Ver ADR-009.",
        "properties": {
          "code": {
            "maxLength": 256,
            "minLength": 16,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "ExchangeIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefreshIn": {
        "additionalProperties": false,
        "properties": {
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshIn",
        "type": "object"
      },
      "Rol": {
        "description": "Los 2 roles del PRD (RBAC de US-403).",
        "enum": [
          "ciudadano",
          "analista"
        ],
        "title": "Rol",
        "type": "string"
      },
      "TokenPair": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "default": 900,
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenPair",
        "type": "object"
      },
      "UserOut": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Rol"
          },
          "sub": {
            "title": "Sub",
            "type": "string"
          }
        },
        "required": [
          "sub",
          "email",
          "role"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Contrato REST de FARO (Escuela como Sensor Social). Fuente de verdad: vault/03_Architecture/API_Specification.md. Este servicio publica el OpenAPI que consumen los mocks de las Células 2 y 3 (US-401).",
    "title": "FARO API — Contrato v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/callback": {
      "get": {
        "description": "Callback de Google: valida el `state`, canjea el `code`, resuelve el rol y entrega la sesion.\n\nDos salidas, segun si `/auth/login` llevaba `redirect` (el `state` firmado lo recuerda):\n\n- **Con `redirect`** (FARO Web): **302** al front con `?code_faro=<codigo de un solo uso>`. Los\n  tokens NO viajan por la URL -- ver ADR-009 y `security/codigos_login.py`.\n- **Sin `redirect`**: el `TokenPair` como JSON, igual que antes (clientes no-navegador y pruebas).",
        "operationId": "callback_api_v1_auth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Callback",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/exchange": {
      "post": {
        "description": "Canjea el codigo de un solo uso de `?code_faro=` por el par de JWT (US-405, ADR-009).\n\nLo llama el **servidor** del frontend, no el navegador: por eso los tokens viajan en el cuerpo\nde la respuesta y nunca por la URL. El codigo se consume en el primer canje; un segundo intento\ncon el mismo codigo (o uno expirado, inventado o ya usado) responde **401**, sin distinguir\nentre los casos.\n\nEl rol se **re-resuelve** aqui con la politica vigente, no se confia en el que quedo guardado:\nsi `ANALISTA_EMAILS` cambio entre el callback y el canje, manda la politica actual.",
        "operationId": "exchange_api_v1_auth_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/login": {
      "get": {
        "description": "Inicia OAuth2 con Google redirigiendo a la pantalla de consentimiento.\n\nGenera el `state` anti-CSRF, lo manda a Google en la URL y guarda el mismo valor en una cookie\n`HttpOnly` para poder compararlos al volver.\n\n`redirect` (opcional, US-405) es la URL de FARO Web a la que volver con el codigo de un solo\nuso. **Se valida contra una allowlist**: sin eso tendriamos un open redirect, y uno dentro del\nflujo de login es el vehiculo clasico para desviar el codigo de autorizacion a un tercero.\nSin `redirect`, el callback responde el `TokenPair` como JSON (clientes que no son navegador).",
        "operationId": "login_api_v1_auth_login_get",
        "parameters": [
          {
            "in": "query",
            "name": "redirect",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Redirect"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Devuelve el usuario autenticado (requiere access token válido; 401 en caso contrario).",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Me",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Canjea un refresh token válido por un par nuevo, re-resolviendo el rol con la política vigente.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "Autenticación"
        ]
      }
    },
...
  }
}